only publish to gh-pages if on master branch

+++ _deploy.sh
@@ -1,14 +1,18 @@
-# configure your name and email if you have not done so
-git config --global user.email "will@bowdit.ch"
-git config --global user.name "willbowditch"
+if [[ $TRAVIS_BRANCH == 'master' ]]
+  # configure your name and email if you have not done so
+  git config --global user.email "will@bowdit.ch"
+  git config --global user.name "willbowditch"
 
-# clone the repository to the book-output directory
-git clone -b gh-pages \
-  https://${GITHUB_PAT}@github.com/${TRAVIS_REPO_SLUG}.git \
-  book-output
-cd book-output
-git rm -rf *
-cp -r ../docs/* ./
-git add --all *
-git commit -m "Update the book"
-git push -q origin gh-pages
+  # clone the repository to the book-output directory
+  git clone -b gh-pages \
+    https://${GITHUB_PAT}@github.com/${TRAVIS_REPO_SLUG}.git \
+    book-output
+  cd book-output
+  git rm -rf *
+  cp -r ../docs/* ./
+  git add --all *
+  git commit -m "Update the book"
+  git push -q origin gh-pages
+else
+  echo "Not publishing as not in MASTER branch"
+fi
